Consolidate sc2_layout.c into stb_sc2layout.h as single-header library
- All types (sc2BaseFrame_t, sc2Frame_t, sc2Layout_t, enums, flags) now
  live in stb_sc2layout.h — the canonical source
- Full XML parser, template resolution, anchor mapping, and frame
  flattening moved into #ifdef STB_SC2LAYOUT_IMPLEMENTATION section
- Two-guard pattern: types guard (always) + impl guard (when defined)
  so unity builds with mixed include orders compile correctly
- sc2_layout.h reduced to thin wrapper (#include stb_sc2layout.h)
- sc2_layout.c reduced to thin trigger (#define IMPLEMENTATION + include)
- hud.c uses SC2_LayoutGetFrames() API instead of direct sc2_layout access
- Expose SC2_MapFrameType() for hud.c fallback builder
